// src/server/api/models/banking/models/CreditControl.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Branch } from './Branch';
import type { CreditControlStatus } from './CreditControlStatus';
import type { LocalDateTime } from './LocalDateTime';
import type { UUID } from './UUID';

export type CreditControl = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The status of the credit control rule
     */
    status: CreditControlStatus;
    /**
     * A unique code generated by the system
     */
    code: string;
    /**
     * The credit rule description
     */
    label: string;
    /**
     * The exposure currency
     */
    currency: string;
    /**
     * The branch that the rule is attached to
     */
    branch: Branch;
    /**
     * The maximum exposure amount for any client
     */
    maxClientExposure: number;
    /**
     * The minimum days in arrears before a write off can be approved
     */
    minDaysInArrearsBeforeWriteOff: number;
    /**
     * The maximum number of accounts that can be attached to an account group
     */
    maxAccountGroupSize: number;
};